Adiciona senha para deletar card ou deck

+++ controllers/deck.js
@@ -33,8 +33,12 @@ const create = rescue(async (req, res) => {
 
 const remove = rescue(async (req, res) => {
   const { id } = req.params;
+  const { password } = req.body;
   const isValid = await deckServices.canRemove(id);
   if (isValid.message) return res.status(isValid.status).json({ message: isValid.message });
+
+  if(!password || password !== 'aiiipapaiii') return res
+  .status(401).json({ message: 'Invalid Password. Only the administrator can remove a deck.' })
   
   const deleted = await Deck.findOne({ where: { id } });
   await Deck.destroy({ where: { id } });
